tune UI
- back button to hide bottom sheet
- status bar color matching top app tool bar
- cleanup code
